backend/requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.0,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
pytest>=7.0,<9.0
pytest-django>=4.0,<5.0
